+++ odoo/local-src/specific_purchase/models/purchase_order.py
@@ -96,11 +96,6 @@ class PurchaseOrder(models.Model):
             else:
                 order.last_date_done = False
 
-    @api.onchange('supplier_promotion_allowed')
-    def onchange_supplier_promotion_allowed(self):
-        for line in self.order_line:
-            line.compute_promotion_supplier()
-
     @api.model
     def delay_update_for_open_po(self):
         """ Delay jobs to update values on all open purchase orders """
@@ -266,7 +261,7 @@ class PurchaseOrderLine(models.Model):
                 uom_id=self.product_uom)
             self.promotion_supplier = (
                 seller.discount_purchase
-                if seller and self.order_id.supplier_promotion_allowed
+                if seller
                 else 0.0
             )
         else: