+++ odoo/local-src/specific_purchase/tests/test_purchase_order.py
@@ -162,8 +162,6 @@ class TestPurchaseOrder(common.TransactionCase):
 
         line = purchase.order_line
 
-        self.assertFalse(supplier.supplier_promotion_purchase_allowed)
-        self.assertFalse(purchase.supplier_promotion_allowed)
         self.assertEquals(line.promotion_supplier, 0)
         self.assertEquals(line.price_unit_base, 100)
         self.assertEquals(line.price_unit, 100)
@@ -174,18 +172,6 @@ class TestPurchaseOrder(common.TransactionCase):
         })
         line.compute_promotion_supplier()
 
-        self.assertFalse(supplier.supplier_promotion_purchase_allowed)
-        self.assertFalse(purchase.supplier_promotion_allowed)
-        self.assertEquals(line.promotion_supplier, 0)
-        self.assertEquals(line.price_unit_base, 100)
-        self.assertEquals(line.price_unit, 100)
-        self.assertEquals(purchase.amount_untaxed, 100)
-
-        supplier.supplier_promotion_purchase_allowed = True
-        line.compute_promotion_supplier()
-
-        self.assertTrue(supplier.supplier_promotion_purchase_allowed)
-        self.assertTrue(purchase.supplier_promotion_allowed)
         self.assertEquals(line.promotion_supplier, 10)
         self.assertEquals(line.price_unit_base, 100)
         self.assertEquals(line.price_unit, 90)