+++ HISTORY.rst
@@ -26,6 +26,7 @@ Unreleased
   * Enable the flag "Expects a Chart of Accounts" on the company
   * Install the module account_chart_update (OCA - account-financial-tools) to create new taxes later (if needed)
   * Create a song to create new Antibiotic taxes
+* ALCYN-464: Remove the supplier_promotion_allowed flag to the purchase order
 
 **Bugfixes**